refactor(dot11): split field and enum tests into tests fields module
